refactor: extract size into props for video customized prop rendering
