+++ app/step/3/components/Letter/Canvas.jsx
@@ -2,25 +2,6 @@
 
 import { useEffect, useState } from "react"
 
-const useCanvasSize = ({ containerRef }) => {
-    const [size, setSize] = useState(0)
-
-    useEffect(() => {
-        setSize(containerRef.current.offsetWidth)
-
-        const resizeHandler = () => {
-            setSize(containerRef.current.offsetWidth)
-        }
-
-        window.addEventListener("resize", resizeHandler)
-        return () => window.removeEventListener("resize", resizeHandler)
-    }, [containerRef])
-
-    return {
-        size: `${size}px`,
-    }
-}
-
 const useMounted = () => {
     const [isMounted, setIsMounted] = useState(false)
 
@@ -31,12 +12,11 @@ const useMounted = () => {
     return { isMounted }
 }
 
-const Canvas = ({ children, containerRef }) => {
+const Canvas = ({ children, size }) => {
     const { isMounted } = useMounted()
-    const { size } = useCanvasSize({ containerRef })
     const canvasSize = {
-        width: size,
-        height: size,
+        width: `${size}px`,
+        height: `${size}px`,
     }
 
     return (
