+++ app/step/3/components/Letter/index.jsx
@@ -1,8 +1,27 @@
 "use client"
 
-import { useRef, useState } from "react"
+import { useEffect, useRef, useState } from "react"
 import { Canvas } from "./Canvas"
 import { EditorToolBar, EditorToolBarSelector } from "./Editor"
+
+const useCanvasSize = ({ containerRef }) => {
+    const [size, setSize] = useState(0)
+
+    useEffect(() => {
+        setSize(containerRef.current.offsetWidth)
+
+        const resizeHandler = () => {
+            setSize(containerRef.current.offsetWidth)
+        }
+
+        window.addEventListener("resize", resizeHandler)
+        return () => window.removeEventListener("resize", resizeHandler)
+    }, [containerRef])
+
+    return {
+        size,
+    }
+}
 
 const Letter = () => {
     const [selectedOption, setSelectedOption] = useState("")
@@ -12,15 +31,14 @@ const Letter = () => {
     }
 
     const letterContainerRef = useRef()
+    const { size } = useCanvasSize({ containerRef: letterContainerRef })
 
     return (
         <div
             ref={letterContainerRef}
             className="flex items-center justify-center h-fit w-full relative"
         >
-            <Canvas containerRef={letterContainerRef}>
-                이곳에 그려봅시다!
-            </Canvas>
+            <Canvas size={size}>이곳에 그려봅시다!</Canvas>
 
             <EditorToolBarSelector
                 handleClickOptions={handleClickOptions}
